requirements.txt
certifi==2018.11.29
chardet==3.0.4
dateparser==0.7.1
future==0.17.1
idna==2.8
oauthlib==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-twitter==3.5
pytz==2018.9
regex==2019.2.7
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.2.18
tzlocal==1.5.1
urllib3==1.24.1
